+++ src/pages/LoadUnload/components/MapLayers/MapLayers.tsx
@@ -1,8 +1,9 @@
 import { BaseLayer } from '@amsterdam/arm-core'
+import { TileLayer } from '@amsterdam/react-maps'
 
 import { AddressMarker } from '../../../../shared/components/MapLayers/AddressMarker'
 import { HighlightedFeatureLayer } from '../../../../shared/components/HighlightedFeatureLayer'
-import { topoBlackWhite } from '../../../../shared/map/mapLayers'
+import { oneWayArrows, topoBlackWhite } from '../../../../shared/map/mapLayers'
 
 import { useLoadUnloadMapContext } from '../../contexts/MapContext'
 import { useLoadUnloadPageContext } from '../../contexts/PageContext'
@@ -30,6 +31,8 @@ export const LoadUnloadMapLayers = () => {
       <LoadUnloadRoadSectionsLoadUnloadLayer />
 
       <LoadUnloadLoadUnloadSpacesLayer />
+
+      <TileLayer options={oneWayArrows.options} args={[oneWayArrows.url]} />
 
       <BaseLayer
         baseLayer={topoBlackWhite.url}