+++ src/shared/components/Modal/index.ts
@@ -1,0 +1,10 @@
+export { DataModal } from './DataModal'
+
+export const getMailtoLink = (
+  recipient: string,
+  subject: string,
+  body: string
+) =>
+  `mailto:${recipient}?subject=${encodeURIComponent(
+    subject
+  )}&body=${encodeURIComponent(body)}`