+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Load-Unload: address search
 - Load-Unload: road sections with load-unload information
 - Load-Unload: one-way traffic indicators
+- Data menu item: provides an overview of the used data on a page
 
 ### Changed
 - Load-Unload: increase max zoom level to 21